fix: fix unit tests and source code compilation issues
- Fixed ConfigLoaderTest.testGetBeforeLoad by adding reset() call
- Fixed MappingGenerator boost field not being added to output
- Fixed query-service source code issues:
  - SearchService.java: removed Kotlin-style import aliases
  - HotRecall.java, KeywordRecall.java, VectorRecall.java: fixed Double to Float conversion
  - HotRecall.java, KeywordRecall.java: added FieldValue wrapper for term queries
  - RecallEngine.java: made strategy variable effectively final for lambda capture
  - RecallEngine.java: improved exception handling for parallel recall
- Fixed test compilation issues:
  - RecallEngineTest: corrected VectorConfig reference path
  - VectorizationQueueTest: cast int to long for poll() method
  - VectorizationServiceTest: fixed missing closing parenthesis
  - DataProcessorTest: added missing List import
- Configured Mockito to use subclass mock maker for Java 25 compatibility
- Added surefire JVM args for Java module access

Test Results:
- common: 5/5 tests passing
- config-repo: 17/17 tests passing
- query-service: 30/30 tests passing
- data-sync: tests blocked by pre-existing source code compilation errors

+++ services/query-service/src/main/java/com/search/query/recall/VectorRecall.java
@@ -61,7 +61,8 @@ public class VectorRecall {
             );
 
             return response.hits().hits().stream()
-                    .map(hit -> new RecallResult(hit.id(), hit.score() != null ? hit.score() : 0.0f, "vector"))
+                    .map(hit -> new RecallResult(hit.id(),
+                            hit.score() != null ? hit.score().floatValue() : 0.0f, "vector"))
                     .collect(Collectors.toList());
 
         } catch (Exception e) {
@@ -102,7 +103,8 @@ public class VectorRecall {
             );
 
             return response.hits().hits().stream()
-                    .map(hit -> new RecallResult(hit.id(), hit.score() != null ? hit.score() : 0.0f, "vector"))
+                    .map(hit -> new RecallResult(hit.id(),
+                            hit.score() != null ? hit.score().floatValue() : 0.0f, "vector"))
                     .collect(Collectors.toList());
 
         } catch (Exception e) {
